feat: send JWT token over when making fetch requests (#11)
### tl;dr

- [x] Adds a custom fetch client

### Description

I decided to go with a custom hook that will allow us to get a custom `fetch` client. The hook allows us to use the `useAuth` hook which is super helpful for appending the JWT for users. If we need to make a network request outside of a React component, we can just use regular ol' `fetch`.